fix: persist Treasure Hunt assets across navigation
